📑 :: Dataclass 네이밍 변경 및 파일구조 변경
