--- data/src/main/java/com/semicolon/data/remote/response/SearchSchool.kt
@@ -1,7 +1,0 @@
-package com.semicolon.data.remote.response
-
-import com.google.gson.annotations.SerializedName
-
-data class SearchSchool(
-    @SerializedName("school_list") val school: List<SchoolList>
-)
