--- data/src/main/java/com/semicolon/data/remote/response/SearchUser.kt
@@ -1,7 +1,0 @@
-package com.semicolon.data.remote.response
-
-import com.google.gson.annotations.SerializedName
-
-data class SearchUser(
-    @SerializedName("user_list") val userList: List<UserList>
-)
