+++ data/src/main/java/com/semicolon/data/remote/response/ranks/inquiryRank/school/SchoolRank.kt
@@ -1,12 +1,11 @@
-package com.semicolon.data.remote.response
+package com.semicolon.data.remote.response.ranks.inquiryRank.school
 
 import com.google.gson.annotations.SerializedName
 
-//search school
-data class SchoolList(
+//check school
+data class SchoolRank(
     @SerializedName("agency_code") val agencyCode: String,
     @SerializedName("logo_image_url") val logoImageUrl: String,
     @SerializedName("name") val name: String,
-    @SerializedName("rank") val rank: Int,
     @SerializedName("walk_count") val walkCount: Int
 )
