+++ data/src/main/java/com/semicolon/data/remote/response/ranks/inquiryRank/school/SchoolRankResponse.kt
@@ -1,0 +1,8 @@
+package com.semicolon.data.remote.response.ranks.inquiryRank.school
+
+import com.google.gson.annotations.SerializedName
+
+data class SchoolRankResponse(
+    @SerializedName("my_school_rank") val mySchoolRank: MySchoolRank,
+    @SerializedName("school_list") val schoolRank: List<SchoolRank>
+)
