+++ data/src/main/java/com/semicolon/data/remote/response/ranks/inquiryRank/user/MyRank.kt
@@ -1,10 +1,11 @@
-package com.semicolon.data.remote.response
+package com.semicolon.data.remote.response.ranks.inquiryRank.user
 
 import com.google.gson.annotations.SerializedName
 
-data class UserRankList(
+data class MyRank(
     @SerializedName("name") val name: String,
     @SerializedName("profile_image_url") val profileImageUrl: String,
+    @SerializedName("rank") val rank: Int,
     @SerializedName("user_id") val userId: Int,
     @SerializedName("walk_count") val walkCount: Int
 )
