+++ data/src/main/java/com/semicolon/data/remote/response/ranks/inquiryRank/user/UserRank.kt
@@ -1,11 +1,10 @@
-package com.semicolon.data.remote.response
+package com.semicolon.data.remote.response.ranks.inquiryRank.user
 
 import com.google.gson.annotations.SerializedName
 
-data class MyRank(
+data class UserRank(
     @SerializedName("name") val name: String,
     @SerializedName("profile_image_url") val profileImageUrl: String,
-    @SerializedName("rank") val rank: Int,
     @SerializedName("user_id") val userId: Int,
     @SerializedName("walk_count") val walkCount: Int
 )
