+++ data/src/main/java/com/semicolon/data/remote/response/ranks/search/user/SearchUserResponse.kt
@@ -1,0 +1,7 @@
+package com.semicolon.data.remote.response.ranks.search.user
+
+import com.google.gson.annotations.SerializedName
+
+data class SearchUserResponse(
+    @SerializedName("user_list") val users: List<UserInfo>
+)
